scripts/podio/requirements.txt
requests>=2.28
python-dotenv>=1.0
openai>=1.0
